Rename desktop_main.rs to cli.rs
